+++ tests/testthat/test-python-iterators.R
@@ -22,3 +22,37 @@ test_that("Iterators are drained of their values by iteration", {
   b <- iterate(iter)
   expect_length(b, 0)
 })
+
+
+test_that("infinite iterators can be accessed with iter_next", {
+  skip_if_no_python()
+  
+  # create an infinite generator
+  main <- py_run_string("
+def infinite_generator():
+  n = 0
+  while True:
+    yield n
+    n += 1
+")
+  it <- main$infinite_generator()
+  
+  # iterate and stop when i is 10
+  while(TRUE) {
+    i <- iter_next(it)
+    if (i == 10) break
+  }
+  expect_equal(i, 10)
+})
+
+test_that("iter_next returns sentinel value when it completes", {
+  skip_if_no_python()
+  iter <- test$makeIterator(c(1:5))
+  while (TRUE) {
+    item <- iter_next(iter, NA)
+    if (is.na(item))
+      break
+  }
+  expect_equal(item, NA)
+})
+